解决List中类型错误的问题

+++ Scripts/SharpScripts/Editor/DotEditor/EGUI/FieldDrawer/Drawers/IntListDrawer.cs
@@ -22,7 +22,7 @@ namespace DotEditor.EGUI.FieldDrawer
             valueList = (List<int>)fieldInfo.GetValue(data);
             if (valueList != null)
             {
-                rList = new ReorderableList(valueList, typeof(float), true, true, true, true);
+                rList = new ReorderableList(valueList, typeof(int), true, true, true, true);
                 rList.drawHeaderCallback = (rect) =>
                 {
                     EditorGUI.LabelField(rect, nameContent, EditorStyles.boldLabel);
